requirements.txt
# Replicate API
replicate>=0.22.0

# HTTP and Image Processing
requests>=2.31.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Configuration and Utilities
PyYAML>=6.0
numpy>=1.24.0
tqdm>=4.65.0

# Character Analysis
face-recognition>=1.3.0
mediapipe>=0.10.0
scikit-image>=0.21.0

# Machine Learning (for advanced features)
torch>=2.0.0
torchvision>=0.15.0

# Async support
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# Data validation and CLI
pydantic>=2.0.0
click>=8.1.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: Advanced image processing
scikit-learn>=1.3.0 